chore: move auth to its own router
