+++ packages/botonic-api/src/rest/routes/users.ts
@@ -176,14 +176,4 @@ router
     }
   })
 
-router.route('/auth').post(async (req: any, res) => {
-  // TODO: Restrict access to this endpoint
-  const { userId } = req.body
-  const PRIVATE_KEY = 'shhhhhh'
-  const token = sign({ userId }, PRIVATE_KEY, {
-    algorithm: 'HS256',
-    expiresIn: '24h',
-  })
-  res.json({ token })
-})
 export default router
